Refactor Lion optimizer experiment
