--- src/model/openthaigpt_pretraining_model/optimizers/lion/constants.py
@@ -1,16 +1,0 @@
-import torch
-
-
-MODEL_NAME = "flax-community/gpt2-base-thai"
-BOS_TOKEN = "<|startoftext|>"
-EOS_TOKEN = "<|endoftext|>"
-PAD_TOKEN = "<|pad|>"
-DATASET_NAME = "mc4"
-ROTARY_PCT = 0.25
-ROTARY_EMB_BASE = 10000
-
-DTYPE_CHOICE = {
-    "float32": torch.float32,
-    "bfloat16": torch.bfloat16,
-    "float16": torch.float16,
-}
